backend/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
python-dotenv
gunicorn